moved file creation from constructor to write method

+++ src/main/java/com/martensigwart/fakeload/RandomAccessDiskOutputSimulator.java
@@ -20,37 +20,44 @@ public final class RandomAccessDiskOutputSimulator extends DiskOutputSimulator {
         super();
         this.filePath = filePath;
 
-        // Create file
-        File f = new File(this.filePath);
-        if (!f.exists()) {
-            //noinspection ResultOfMethodCallIgnored
-            f.getParentFile().mkdirs();
-            //noinspection ResultOfMethodCallIgnored
-            f.createNewFile();
-        }
 
-        file = new RandomAccessFile(this.filePath, "rws");
     }
 
     @Override
     protected void write(byte[] bytes) throws IOException {
+
+        // Create file if does not exist yet
+        if (file == null) {
+            File f = new File(this.filePath);
+            if (!f.exists()) {
+                //noinspection ResultOfMethodCallIgnored
+                f.getParentFile().mkdirs();
+                //noinspection ResultOfMethodCallIgnored
+                f.createNewFile();
+                log.trace("Created file {}", this.filePath);
+            }
+            file = new RandomAccessFile(this.filePath, "rws");
+        }
+
         file.seek(0);
         file.write(bytes);
     }
 
     @Override
     protected void cleanUp() {
-        try {
-            file.close();
-            File f = new File(filePath);
-            if (f.delete()) {
-                log.trace("File {} successfully deleted", filePath);
-            } else {
-                log.warn("File {} could not be deleted", filePath);
+        if (file != null) {
+            try {
+                file.close();
+                File f = new File(filePath);
+                if (f.delete()) {
+                    log.trace("File {} successfully deleted", filePath);
+                } else {
+                    log.warn("File {} could not be deleted", filePath);
+                }
+            } catch (IOException e) {
+                log.error("Failed to close file: {}", e.getMessage());
+                e.printStackTrace();
             }
-        } catch (IOException e) {
-            log.error("Failed to close file: {}", e.getMessage());
-            e.printStackTrace();
         }
     }
 }
